requirements.txt
linkedin-api
python-dotenv
fuzzywuzzy
python-Levenshtein
flask
flask_cors
supabase
ollama